+++ app/decorators/image_with_derivatives.rb
@@ -1,4 +1,5 @@
 class ImageWithDerivatives < SimpleDelegator
+  delegate :derivative_decorator_class, :runner_list, :to => :derivative_injector
 
   def save
     derivative_asset.save
@@ -19,18 +20,22 @@ class ImageWithDerivatives < SimpleDelegator
     self.pyramidal_path = path
   end
 
+  def derivative_injector
+    OregonDigital.derivative_injector
+  end
+
   private
 
   def derivative_asset
-    @derivative_asset ||= AssetWithDerivatives.new(__getobj__, runners)
+    @derivative_asset ||= derivative_decorator_class.new(__getobj__, runners)
   end
 
   def runners
-    @runners ||= OregonDigital::Derivatives::Runners::RunnerList.new(
+    @runners ||= runner_list.new(
       [
-        injector.thumbnail_runner(id),
-        injector.medium_runner(id),
-        injector.pyramidal_runner(id)
+        derivative_injector.thumbnail_runner(id),
+        derivative_injector.medium_runner(id),
+        derivative_injector.pyramidal_runner(id)
       ]
     )
   end
